feat: Implement comprehensive Payroll module with administration and employee features, and enhance Personnel management with new profile sections and wizard steps.

+++ Backend/HRMS/HRMS.Application/DTOs/Payroll/LoanInstallmentDto.cs
@@ -5,7 +5,7 @@ namespace HRMS.Application.DTOs.Payroll;
 /// </summary>
 public class LoanInstallmentDto
 {
-    public int InstallmentId { get; set; }
+    public long InstallmentId { get; set; }
     public int LoanId { get; set; }
     public int EmployeeId { get; set; }
     public string? EmployeeName { get; set; }
